allow float as timestamp

+++ trace/tef.py
@@ -23,7 +23,7 @@ class Tracer(object):
 
 
 class TraceEvent(ABC):
-    def __init__(self, name: str, ts: int, pid: int = 0, tid: int = 0, **kwargs):
+    def __init__(self, name: str, ts: float, pid: int = 0, tid: int = 0, **kwargs):
         self.name = name
         self.ts = ts * 1e6  # unit in trace event format is micro-second
         self.pid = pid
